+++ src/main/java/com/glucoseguardian/webbackend/storage/entity/Glicemia.java
@@ -9,8 +9,8 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import java.io.Serializable;
+import java.sql.Date;
 import java.sql.Time;
-import java.util.Date;
 import java.util.Objects;
 
 /**
@@ -22,7 +22,7 @@ public class Glicemia implements Serializable {
   @Id
   @GeneratedValue(strategy = GenerationType.AUTO)
   @Column(columnDefinition = "UNSIGNED INT", nullable = false)
-  private int id;
+  private long id;
   @Column(nullable = false)
   private int livelloGlucosio;
   @Column(nullable = false)
@@ -34,11 +34,11 @@ public class Glicemia implements Serializable {
   @JoinColumn(name = "paziente")
   private Paziente paziente;
 
-  public int getId() {
+  public long getId() {
     return id;
   }
 
-  public void setId(int id) {
+  public void setId(long id) {
     this.id = id;
   }
