+++ src/main/java/com/glucoseguardian/webbackend/storage/entity/Terapia.java
@@ -12,7 +12,7 @@ import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.OneToOne;
 import java.io.Serializable;
-import java.util.Date;
+import java.sql.Date;
 import java.util.List;
 import java.util.Objects;
 
@@ -25,7 +25,7 @@ public class Terapia implements Serializable {
   @Id
   @GeneratedValue(strategy = GenerationType.AUTO)
   @Column(columnDefinition = "UNSIGNED INT", nullable = false)
-  private int id;
+  private long id;
   @Column
   private Date dataInizio;
   @OneToMany(mappedBy = "id", fetch = FetchType.LAZY, cascade = CascadeType.ALL)
@@ -43,11 +43,11 @@ public class Terapia implements Serializable {
   private Paziente paziente;
 
 
-  public int getId() {
+  public long getId() {
     return id;
   }
 
-  public void setId(int id) {
+  public void setId(long id) {
     this.id = id;
   }
